fix: mis-use of serverMiddleware
